requirements.txt
Django>2.0
djangorestframework>=3.8
factory-boy
coreapi-cli
django-cors-headers
Pillow
django-markdownx
dry_rest_permissions
django-admin-sortable2
pymdown-extensions
psycopg2
dj-database-url
django-heroku
gunicorn
django-storages
boto3
whitenoise
celery[redis]
django-countries
